Hack to include a 4th axis (moving pins around)

+++ grbl/planner.h
@@ -37,7 +37,7 @@
 typedef struct {
   // Fields used by the bresenham algorithm for tracing the line
   // NOTE: Used by stepper algorithm to execute the block correctly. Do not alter these values.
-  uint8_t direction_bits;    // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
+  uint16_t direction_bits;    // The direction bit set for this block (refers to *_DIRECTION_BIT in config.h)
   uint32_t steps[N_AXIS];    // Step count along each axis
   uint32_t step_event_count; // The maximum step axis count and number of steps required to complete this block.
